+++ src/navigation/MainAppNav.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { TouchableOpacity } from 'react-native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { NavigationContainer, useNavigation } from '@react-navigation/native';
-import Homepage from '../screens/Homepage';
+import { useNavigation } from '@react-navigation/native';
+import HomeScreen from '../screens/HomeScreen';
 import HabitScreen from '../screens/HabitScreen';
 import CreateHabit from '../screens/CreateHabit';
 import StartHabitCreation from '../screens/StartHabitCreation';
@@ -70,7 +70,7 @@ const MainAppStack = () => {
                             </TouchableOpacity>
                         ),
                     })}
-                    component={Homepage}
+                    component={HomeScreen}
                 />
                 <Stack.Screen
                     name="StartHabitCreation"
